"Understanding LSTM Networks" by colah
Saw this blog post in the comments of the so called rnn-effectiveness paper and wanted to peek for a small understanding of what is LSTM. It turns out this post is very good for understanding the difference between RNN and LSTMs. I can say it very exceeded my expectations. Excited to read a bit about attention before jumping to "Attention is all you need" article.