"Attention and Augmented Recurrent Neural Networks" by Olah & Carter
This was an interval reading for me that introduced the attention and what can be done using attentions in RNN. It briefly explained some examples of usage of attention and how attention works. The main idea is we want the computation criteria to be differentiable so the backpropagation can be done. And we do this by adding attention so the discrete data can be differentiable.