"Show, Attend and Tell: Neural Image Caption Generation with Visual Attention" by Xu et al.
This was the hardest reading i ever had so far. The general language, explanation of the model and the abundance of the prerequisites were giving me hard time. I didn't really understood the paper but got an idea of how attention works and whats the difference between "hard" and "soft" attention. I will definitely read this paper once again when i dive deeper into the context.